Correct the 6G chan to freq conversion change (#4)
* Correct the 6G chan to freq conversion change

Reason for change: The 6G chan to freq conversion change is incorrect
and not required. The current change will not work when CONFIG_DRIVER_BRCM
flag is not defined. Also, the change present in Upstream will work
for all the cases including the CONFIG_DRIVER_BRCM case, thus removing the change.

Test Procedure: Tested 6G AP bringup and connection.

* Remove unnecessary printf in eloop.c